+++ docs/changelog.md
@@ -5,6 +5,7 @@ This changelog documents user-relevant changes to the GitHub runner charm.
 ## 2026-02-27
 
 - Fix a bug where image labels (base,arch) are not propagated to the planner charm.
+- Fix a bug where max_total_virtual_machines is not propagated to the github-runner-manager when used with the planner.
 
 ## 2026-02-25